+++ game.py
@@ -6,12 +6,18 @@
 # Desrcription: A set of dictionaries for a Clue RPG
 
 """ A dictionary of all the characters/ suspects in the case"""
-characters = {"Evelyn Finch": "",
-             "George Brown": "",
-             "Madeline Brown": "",
-             "Stephen Turner": "",
-             "Chantal Steele": "",
-             "Justin Jones": ""
+characters = {"Evelyn Finch":
+                {"Age": "24", "Job": "Interior Designer", "City": "Toronto" },
+             "George Brown":
+                {"Age": "33", "Job": "Doctor", "City": "Chicago"},
+             "Madeline Brown":
+                {"Age": "30", "Job": "Mechanic", "City": "Chicago"},
+             "Stephen Turner":
+                {"Age": "45", "Job": "Butcher", "City": "New York"},
+             "Chantal Steele":
+                {"Age": "21", "Job": "News Reporter", "City": "Vancouver"},
+             "Justin Jones":
+                {"Age": "54", "Job": "Plumber", "City": "Winnipeg"},
              }
 
 """ A dictionary of all of the locations the murder may have taken place"""
@@ -22,3 +28,7 @@ locations = {"library": "",
             "bedroom": "",
             "office": ""
             }
+
+"""Print instructions for the character information"""
+for key in characters:
+    print("Information on ")